refactor: 💡 rename class name
