+++ src/pages/contact/comonents/ConfirmButton/index.module.scss
@@ -1,18 +1,7 @@
 @import '@/styles/mixin';
 @import '@/styles/variables';
 
-.inputWrapper {
-  display: flex;
-  flex-direction: column;
-  row-gap: 10px;
-  width: 200px;
-
-  @include lg {
-    width: 100%;
-  }
-}
-
-.submit {
+.button {
   width: linear-clamp(200, 500);
   padding: 5px;
   color: $colorWhite;
