+++ src/pages/contact/comonents/RadioGroup/index.module.scss
@@ -1,7 +1,7 @@
 @import '@/styles/mixin';
 @import '@/styles/variables';
 
-.checkBoxContainer {
+.container {
   display: flex;
   flex-wrap: wrap;
   gap: 10px;
@@ -9,7 +9,7 @@
   margin-top: 30px;
 }
 
-.checkBoxLabel {
+.checkBoxItem {
   display: flex;
   column-gap: 4px;
   align-items: center;
