+++ src/pages/contact/comonents/TextAriaFiled/index.module.scss
@@ -1,8 +1,22 @@
 @import '@/styles/mixin';
 @import '@/styles/variables';
 
+.container {
+  display: flex;
+  flex-direction: column;
+  row-gap: 10px;
+  width: 200px;
+
+  @include lg {
+    width: 100%;
+  }
+}
+
 .textArea {
+  width: 100%;
   max-width: linear-clamp(200, 500);
+  min-height: linear-clamp(300, 500);
   padding: 10px;
+  resize: vertical;
   background-color: $colorWhite;
 }
