+++ src/pages/contact/comonents/TextField/index.module.scss
@@ -1,7 +1,7 @@
 @import '@/styles/mixin';
 @import '@/styles/variables';
 
-.inputWrapper {
+.container {
   display: flex;
   flex-direction: column;
   row-gap: 10px;
@@ -12,7 +12,7 @@
   }
 }
 
-.flex {
+.fieldHeader {
   display: flex;
   column-gap: 10px;
   align-items: center;
